code coverage fix for code climate
Moved the unit tests of Item and User classes to Android Tests to make code climate get code coverage.
